Changed All Functionality
My orignal plan was to have people who are developing C++ to use C+ instead of g++ so I made some code that would allow you to input gcc commands but because I cant search directories on computers without some kernel crap Im not going to even bother.
So now the usage is -options <args> [files]. Basicly like how you would set up gcc on the command line.

Also there are alot of bug fixes as well as a hot function that finds the first instance of a string inside of a string. Not too hard in C++ but coding in C is alot more rewarding.
One of the dumbest errors that I had in this version is that I defined true to be 0. Yea that cause a lot of errors. anyway, Assembler is coming along.  yay!